Squashed commit of the following:
commit 662fd7ce2374ed2e6002e4c602037dee8fea27ec
Author: hikhhbro <w1121508831@gmail.com>
Date:   Mon Jan 18 18:16:44 2021 +0800

    添加csv

commit c3752ad83db90785dd0d99e1534f5eaebc60b78c
Author: hikhhbro <w1121508831@gmail.com>
Date:   Mon Jan 18 13:51:35 2021 +0800

    完成底盘测试

commit 77f75150a21d7c51591c35209702406b54545e35
Author: hikhhbro <w1121508831@gmail.com>
Date:   Mon Jan 18 13:40:22 2021 +0800

    修复在角度控制小于5度是无法修正的bug

commit 58c16a740b3b655e89f944696265610c0dd1c65f
Author: hikhhbro <w1121508831@gmail.com>
Date:   Mon Jan 18 11:42:35 2021 +0800

    添加log

+++ src/main/include/BaseModule.h
@@ -8,7 +8,6 @@
 #include <iostream>
 #include "rev/SparkMax.h"
 #include "ctre/Phoenix.h"
-
 #define IS_SECTION(data_,min,max) (  ((data_) > (min) && (data_) < (max)) ?true:false) 
 #define IS_X_SECTION(data_,x)   (((data_) > (-abs(x)) && (data_) < (abs(x))) ?true:false) 
 
@@ -90,13 +89,13 @@ public:
     Falcon();
     ~Falcon();
     float enc_100ms_to_rpm(float);
-    int rpm_to_enc_100ms(float);
-    int mm_to_enc(float);
+    float rpm_to_enc_100ms(float);
+    float mm_to_enc(float);
     double enc_to_mm(int);
     void set_dia(float);
     float d = 100;//mm
     int get_position_error(int target,int real);
-    int mms_to_enc100ms(float mms);
+    float mms_to_enc100ms(float mms);
     float enc100ms_to_mms(int enc);
     int max_enc_100ms = 20000;
     float enc100ms_to_per(float enc){return enc/(float)max_enc_100ms;};
